#91: Incoming FlowFiles no longer get stuck before DeleteMarkLogic
This introduces a new "original" relationship to QueryML/DeleteML that users will need to configure. Any incoming FF will be sent to this relationship. This fixes the bug where the first session.commit() call failed because the incoming FF was not configured to be sent anywhere. 

There is still one leftover failing test which will be addressed in a separate ticket. 

Did a little test cleanup - had a bunch of "teardown()" calls that weren't doing anything.