#94 Fixing failing tests
Added logging to QueryMarkLogic to both help debug the Apply test, and it seems useful in general. 

Moved AddAttribute into a transform that gets deployed via mlDeploy. The test was hanging when installing a transform, writing data, and then running the processor. Does not hang when only writing data and running the processor. 

Removed seemingly useless onScheduled calls in PutMarkLogicDuplicateUriTest that were producing NPEs.