#90: ExtensionCallML now routes errors to failure relationship
In addition, "markLogicErrorMessage" is added as an attribute to the flowfile. 

Also removed AllowableValuesSet class, which had a field that wasn't being used anywhere, and thus the class wasn't doing anything. 

I am thinking that a follow-up ticket will change the other processes that rollback the session to instead route the failed FF to a "failure" relationship with the error message as a new attribute. That seems like a better UX, as it gives the flow developer a way to figure out what to do with the FF.